\name{greene}
\docType{data}
\alias{greene}
\title{sample scrips from \emph{gretl} 'Greene' selection}
\description{sample scrips from \emph{gretl} 'Greene' selection translated as \R character vectors; they can be executed via \code{run_grcli}.}
\usage{greene
# names(greene)
# show_sample(greene$greene7.8)
# run_grcli(greene$greene7.8)}
\format{
  List containing 13 sample scripts (= character vectors) which can be executed via \code{run_grcli}. See entry for \code{ramanathan} data set.
  }
  \source{\emph{gretl} sample scripts}
 \keyword{datasets} 


