\name{gretl}
\docType{data}
\alias{gretl}
\title{sample scrips from \emph{gretl} 'Gretl' selection}
\description{sample scrips from \emph{gretl} 'Gretl selection translated as \R character vectors; they can be executed via \code{run_grcli}.}
\usage{gretl
# names(gretl)
# show_sample(gretl$reprobit)
# run_grcli(gret$reprobit)}
\format{
  List containing 40 sample scripts (= character vectors) which can be executed via \code{run_grcli}. See entry for \code{ramanathan} data set.
  }
  \source{\emph{gretl} sample scripts}
 \keyword{datasets} 


